pyproject.toml
[tool.poetry]
name = "llm-telegram-bot"
version = "0.1.0"
description = "telegram bot for access LLM via chat"
authors = ["Evgeniy Korzunov <korzunov.e.v@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
python-telegram-bot = {extras = ["job-queue"], version = "^21.10"}
anthropic = "0.47.1"
pymongo = {extras = ["srv"], version = "^4.10.1"}
requests = "^2.32.3"
langchain-text-splitters = "^0.3.5"
pdf2txt = "^0.7.14"
telegramify-markdown = "^0.4.2"
openai = "^1.63.2"
pydantic-ai-slim = {extras = ["anthropic", "openai"], version = "^0.0.24"}
tiktoken = "^0.9.0"
aiohttp = "^3.11.12"
aiofiles = "^24.1.0"


[tool.poetry.group.dev.dependencies]
poetry = "1.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
